speech2text/requirements.txt
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
contextlib2==21.6.0
Django==4.2.1
django-ninja==0.21.0
django-ninja-extra==0.18.8
idna==3.4
injector==0.20.1
packaging==23.1
pydantic==1.10.7
replicate==0.8.1
requests==2.30.0
sqlparse==0.4.4
typing_extensions==4.5.0
urllib3==2.0.2
